[NCL-3812] Improve Build Configuration section on Project detail page

+++ ui/app/build-configs/directives/pnc-build-configs-data-table/pncBuildConfigsDataTable.js
@@ -88,10 +88,12 @@
     };
 
     function remove(buildConfig) {
-      $q.when($ctrl.onRemove()(buildConfig)).then(() => {
-        $ctrl.refreshBuildConfigs()($ctrl.page.data.filter(bc => bc.id !== buildConfig.id));
-        $ctrl.filterPage.refresh();
-      });
+      if ($ctrl.onRemove){
+        $q.when($ctrl.onRemove()(buildConfig)).then(() => {
+          $ctrl.refreshBuildConfigs()($ctrl.page.data.filter(bc => bc.id !== buildConfig.id));
+          $ctrl.filterPage.refresh();
+        });
+      }
     }
 
     function edit() {
